requirements.txt
flask==2.0.1
pymongo==3.11.4
Flask-RESTful==0.3.9
gunicorn==20.1.0
pytest-flask==1.2.0